pyproject.toml
[project]
name = "csfloat-sales-parser"
version = "0.1.0"
description = "Parser for CSFloat latest sales"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "selenium>=4.43.0",
    "sqlalchemy>=2.0.49",
]
